intrinsic_dimension/requirements.txt
numpy
h5py
IPython
colorama
scikit-learn
tensorflow-gpu
keras
